Passes dummy events to programatic open and close methods
Originally I was passing in an object with a null 'delegateTarget'
attribute because that is what `delegate` appends to the event. I
eventually changed this to a poor man's optional chaining via
`event && event.delegateTarget` and so removed this "dummy" event.

After using this library a little bit more, I had instances where
I wanted to call `event.preventDefault()` or access other attributes
on the event or `trigger_node`, but had to ensure the event's existence
first because I also was accessing to programatic `open()` and
`close()` methods.

Having to write `event && event.preventDefault()` was a bit annoying.
So, I figured it would be better to _always_ pass in a _real_ event
that you can safely access its members without having to worry about
its existence in the first place.