Passes in open event to dynamic context calls
This perhaps exposes a better way to create dynamic modals.

Let's say you wanted to create an "interstitial" modal on a link. That
is, when you clicked on a link, rather than it immediately navigating,
it instead would open a modal first. This is common with "third party
links."

Using this library, if you wanted to do something like that, you'd
have to do something like this:

    // <a href="http://example.com" data-modal-trigger>3rd party</a>

    yamodal({
      template: () => `<div>Continue? <a href="">Link</a></div>`,
      beforeInsertIntoDom(modal, trigger, e) {
        e.preventDefault();
        let href = trigger.getAttribute('href');
        let inner_link = modal.querySelector('a');
        inner_link.setAttribute('href', href);
      },
    });

Grabbing the inner link by querying the modal node seems a little
cumbersome. Now, by passing the event to our context function (if it
exists), we can make this a little cleaner

    // <a href="http://example.com" data-modal-trigger>3rd party</a>

    yamodal({
      template: url => `<div>Continue? <a href="${url}">Link</a></div>`,
      context(e) {
        // Can be here or in `beforeInsertIntoDom`
        e.preventDefault();

        let trigger = e.delegateTarget;
        return trigger.getAttribute('href');
      },
    });

Hmm, although, writing that out here, is that cleaner? I don't know.
Maybe I should also be passing in the trigger_node into the context
function. But then that complicates my `createModalNode` function...

May have to let this one marinate a bit.