Passes in trigger_node and event into context
This hopefully makes the previous example a bit easier to work with.

So now, for that "interstitial" example, we can do this:

    // <a href="http://example.com" data-modal-trigger>3rd party</a>

    yamodal({
      template: url => `<div>Continue? <a href="${url}">Link</a></div>`,
      beforeInsertIntoDom(modal, trigger, event){
        event.preventDefault();
      },
      context(trigger) {
        return trigger.getAttribute('href');
      },
    });

With this, I also made some changed with how the `modal_node` is
initially created, as well as the error checking.

Although, now that I think about it again (!!), I think moving the
error checking to this place isn't a good idea, because even though
this adds a specific Error, in the dynamic `context` situation,
there wouldn't be a way to catch the error, since the triggerOpen
is an event handler.

Ehh, maybe its still a good thing. This seems like such an edge case
(that is, your dynamic context changes the template in such a way that
it no longer returns a DOM node).

I'm going to keep it until I see a compelling reason to handle this
differently.